+++ internal/migrate/unstage_contract_test.go
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package migration
+package migrate
 
 import (
 	"testing"
